@@ -42,25 +42,25 @@
     "alex": "alex README.md"
   },
   "devDependencies": {
-    "alex": "^2.0.0",
-    "babel": "^6.5.1",
+    "alex": "^2.0.1",
+    "babel": "^6.5.2",
     "babel-cli": "^6.5.1",
-    "babel-core": "^6.5.1",
-    "babel-eslint": "^4.1.8",
+    "babel-core": "^6.5.2",
+    "babel-eslint": "^5.0.0",
     "babel-jest": "^6.0.1",
-    "babel-loader": "^6.2.2",
+    "babel-loader": "^6.2.3",
     "babel-plugin-react-transform": "^2.0.0",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-react-hmre": "^1.1.0",
     "babel-preset-stage-0": "^6.5.0",
     "css-loader": "^0.23.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "5.0.0",
-    "eslint-plugin-mocha": "^1.1.0",
+    "eslint": "^2.1.0",
+    "eslint-config-airbnb": "5.0.1",
+    "eslint-plugin-mocha": "^2.0.0",
     "eslint-plugin-react": "^3.16.1",
     "jest-cli": "^0.8.2",
-    "jscs": "^2.9.0",
+    "jscs": "^2.10.1",
     "react": "^0.14.7",
     "react-addons-test-utils": "^0.14.7",
     "react-dom": "^0.14.7",
@@ -70,7 +70,7 @@
     "style-loader": "^0.13.0",
     "webpack": "^1.12.13",
     "webpack-dev-middleware": "^1.5.1",
-    "webpack-hot-middleware": "^2.6.4"
+    "webpack-hot-middleware": "^2.7.1"
   },
   "jest": {
     "testRunner": "<rootDir>/node_modules/jest-cli/src/testRunners/jasmine/jasmine2.js",
